requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# FHIR Resources Library
fhir.resources==7.1.0

# Database
pymongo==4.6.0
motor==3.3.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP Client
httpx==0.25.2
aiofiles==23.2.1

# Date/Time
python-dateutil==2.8.2

# Environment Variables
python-dotenv==1.0.0

# Validation
email-validator==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1
